Signal Handling
Removed the for() loop for the base variable because it's exceedingly unnecessary.

Added in signal handling so when timeout is used with the program it exits gracefully, giving a final count.
(to use timeout you do "timeout -s 10 *S* mpiexec -np *P* ./a.out" where *S* is the signal to send and *P* is the number of processes. On ubuntu you send signal 10 for SIGUSR1)

was there anything else I had to do?